Move AudioFrame to its own header
Use AudioFrame from new header
